backend/scripts/fetch_omb_data.py
"""
OMB Data Fetcher for Federal Budget Dashboard

This script will fetch real federal budget data from OMB sources:
- OMB Historical Tables (Excel/CSV format)
- Public Budget Database
- Federal Spending Transparency data

TODO: Implement actual data fetching and parsing
For now, this is a placeholder that documents the data sources.
"""

import os
import requests
import pandas as pd
from typing import Dict, List, Any

# OMB Data Sources
OMB_HISTORICAL_TABLES_URL = "https://www.whitehouse.gov/omb/historical-tables/"
OMB_PUBLIC_BUDGET_DB_URL = "https://www.whitehouse.gov/omb/budget/"

# Specific files we'll need:
OMB_FILES = {
    "outlays_by_function": "https://www.whitehouse.gov/wp-content/uploads/2024/03/hist03z1.xlsx",
    "outlays_by_agency": "https://www.whitehouse.gov/wp-content/uploads/2024/03/hist04z1.xlsx", 
    "budget_authority": "https://www.whitehouse.gov/wp-content/uploads/2024/03/hist31z1.xlsx",
    "function_categories": "https://www.whitehouse.gov/wp-content/uploads/2024/03/hist32z1.xlsx"
}

def fetch_omb_file(url: str, filename: str) -> str:
    """
    Download an OMB file and save locally
    
    Args:
        url: URL to download from
        filename: Local filename to save as
        
    Returns:
        Path to the downloaded file
    """
    print(f"Downloading {filename} from OMB...")
    
    response = requests.get(url)
    response.raise_for_status()
    
    local_path = f"data/{filename}"
    os.makedirs("data", exist_ok=True)
    
    with open(local_path, 'wb') as f:
        f.write(response.content)
    
    print(f"✅ Downloaded {filename}")
    return local_path

def parse_outlays_by_function(file_path: str) -> List[Dict[str, Any]]:
    """
    Parse OMB Historical Table for outlays by function
    
    Args:
        file_path: Path to the Excel file
        
    Returns:
        List of budget records
    """
    # TODO: Implement Excel parsing
    # This is complex because OMB files have:
    # - Multiple sheets
    # - Merged cells
    # - Headers spanning multiple rows
    # - Years as columns (need to transpose)
    
    print(f"Parsing {file_path}...")
    
    # Placeholder implementation
    return [
        {
            "category": "National Defense",
            "function_code": "050",
            "fiscal_year": 2024,
            "amount": 850000000000,
            "data_type": "outlays"
        }
    ]

def parse_outlays_by_agency(file_path: str) -> List[Dict[str, Any]]:
    """
    Parse OMB Historical Table for outlays by agency
    """
    # TODO: Implement
    print(f"Parsing {file_path}...")
    return []

def clean_and_standardize_data(raw_data: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
    """
    Clean and standardize the OMB data for our database format
    
    This includes:
    - Converting function codes to readable categories
    - Handling inflation adjustments
    - Standardizing agency names
    - Converting amounts to integers
    """
    # TODO: Implement data cleaning
    return raw_data

def main():
    """
    Main function to fetch and process OMB data
    """
    print("🏛️  Fetching OMB Federal Budget Data...")
    print("=" * 50)
    
    print("⚠️  This script is a placeholder for future implementation")
    print("Current implementation uses sample data in seed_database.py")
    print()
    print("To implement real OMB data fetching:")
    print("1. Download Excel files from OMB Historical Tables")
    print("2. Parse complex Excel format with pandas/openpyxl")
    print("3. Clean and standardize category names")
    print("4. Map function codes to readable categories")
    print("5. Handle multi-year data")
    print()
    print("OMB Data Sources:")
    for name, url in OMB_FILES.items():
        print(f"  - {name}: {url}")
    
    # Future implementation would do:
    # 1. Download files
    # 2. Parse Excel sheets
    # 3. Clean and standardize
    # 4. Save to database or CSV
    
    print("\n✅ Use seed_database.py for now to populate with sample data")

if __name__ == "__main__":
    main() 